include arrows in style sheet + move editor graph type to separate file
