+++ editor/jvm/src/test/scala/com/flowtick/graphs/MathUtilSpec.scala
@@ -1,8 +1,7 @@
 package com.flowtick.graphs
 
-import com.flowtick.graphs.editor.MathUtil
-import com.flowtick.graphs.editor.MathUtil.{LineSegment, Vector2}
-
+import com.flowtick.graphs.editor.util.MathUtil.{LineSegment, Vector2}
+import com.flowtick.graphs.editor.util.MathUtil
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
