+++ editor/shared/src/main/scala/com/flowtick/graphs/editor/util/MathUtil.scala
@@ -1,4 +1,4 @@
-package com.flowtick.graphs.editor
+package com.flowtick.graphs.editor.util
 
 /**
  * line segment intersection adapted from
